+++ controllers/users.js
@@ -9,8 +9,32 @@ const getAll = async (req, res, next) => {
   }
 };
 
+const login = async (req, res, next) => {
+  try {
+    const jwt = await usersService.login(req.body);
+    if (!jwt) {
+      res.status(401).json({ ok: 'false' });
+    } else {
+      res.status(200).json({ token: jwt });
+    }
+  } catch (e) {
+    next(e);
+  }
+};
+const create = async (req, res, next) => {
+  try {
+    const user = await usersService.create(req.body);
+    res.status(200).json({
+      firstName: user.firstName, lastName: user.lastName, email: user.email
+    });
+  } catch (e) {
+    next(e);
+  }
+};
 
 module.exports = {
   getAll,
-  
+  login,
+  create
+
 };
