+++ routes/auth.js
@@ -2,8 +2,7 @@ const express = require('express');
 
 const router = express.Router();
 
-
-const authController = require('../controllers/auth');
+const authController = require('../controllers/users');
 
 const authMiddleware = require('../middlewares/auth');
