+++ routes/users.js
@@ -2,7 +2,8 @@ const express = require('express');
 
 const router = express.Router();
 const usersController = require('../controllers/users');
+const authMiddleware = require('../middlewares/auth');
 /* GET users listing. */
-router.get('/', usersController.getAll);
+router.get('/', authMiddleware.isAdmin, usersController.getAll);
 
 module.exports = router;
